delete and search updated. delete and serach unit testing added jest config

+++ backend/server/api/services/users/deleteItem.js
@@ -1,15 +1,20 @@
 const SQL = require('sql-template-strings');
-const { query, ErrorException } = require('../helpers');
+const { query, queryOne, ErrorException } = require('../helpers');
 
 const create = async (id) => {
   if (!id) throw new ErrorException('Missing ID', 500);
+  const checkDB = await queryOne(
+    SQL`SELECT "_id", "name" FROM "Users" WHERE "_id"=${id}`
+  );
+  if (!checkDB) throw new ErrorException('No User found', 500);
+
   const statement = SQL`
     DELETE FROM "Users"
       WHERE "_id"=${id}
     `;
   await query(statement);
   return {
-    message: `Deleted User`,
+    message: 'Deleted User',
   };
 };
